fix: make RedBox focusing a little more sane
We want to ensure the RedBox is visible when an error occurs,
but we also want to respect the developer when they close the
RedBox before it's finished symbolicating. Lastly, the RedBox
should auto-hide when the bridge reloads. Note: We can't hide
the RedBox when a syntax error occurs, because syntax errors
require the developer to reload manually.