fix(RCTView): stop needing a visible `backgroundColor` to draw the border
For views that are rounded rectangles without `overflow: "hidden"`, there was a race condition where the `setBackgroundColor:` method would not redraw the border even if all conditions were technically met.

This could easily lead to an incorrect-looking border.

There were two options to fix this: (1) call `setNeedsDisplay` whenever the `backgroundColor` is changed, or (2) draw a transparent border even if it means wasting memory with an empty image.

I went with option 2, because of how rare it is for a view to have rounded corners, but no visible borders nor a background color. Also, it might be expensive to animate the `backgroundColor` if it needs to check the border props every time it changes. Idk, maybe I'm being paranoid. Either way, option 2 was also easiest to implement.